+++ docs/source/news.rst
@@ -4,9 +4,17 @@ What's New
 0.8.0
 ^^^^^
 
+Features:
+
 * ``Message.log`` will log a new message, combining the existing ``Message.new`` and ``Message.write``.
 * ``write_traceback`` and ``writeFailure`` no longer require a ``Logger``; they now default to using the global one.
 * The logs written with ``redirectLogsForTrial`` are now written in JSON format, rather than with ``pformat``.
+
+Bug fixes:
+
+* ``FileDestination`` will now call ``flush()`` on the given file object after writing the log message.
+  Previously log messages would not end up being written out until the file buffer filled up.
+* Each ``Message`` logged outside the context of an action now gets a unique ``task_id``.
 
 
 0.7.0